third_party/blink/web_tests/external/wpt/css/css-animations/parsing/animation-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Default animation value
  assert_equals: expected "none" but got "none 0s ease 0s 1 normal none running"
[FAIL] Property animation value '1s'
  assert_equals: expected "1s" but got "1s ease 0s 1 normal none running none"
[FAIL] Property animation value 'cubic-bezier(0, -2, 1, 3)'
  assert_equals: expected "cubic-bezier(0, -2, 1, 3)" but got "0s cubic-bezier(0, -2, 1, 3) 0s 1 normal none running none"
[FAIL] Property animation value 'ease-in-out'
  assert_equals: expected "ease-in-out" but got "0s ease-in-out 0s 1 normal none running none"
[FAIL] Property animation value '1s -3s'
  assert_equals: expected "1s -3s" but got "1s ease -3s 1 normal none running none"
[FAIL] Property animation value '4'
  assert_equals: expected "4" but got "0s ease 0s 4 normal none running none"
[FAIL] Property animation value 'reverse'
  assert_equals: expected "reverse" but got "0s ease 0s 1 reverse none running none"
[FAIL] Property animation value 'both'
  assert_equals: expected "both" but got "0s ease 0s 1 normal both running none"
[FAIL] Property animation value 'paused'
  assert_equals: expected "paused" but got "0s ease 0s 1 normal none paused none"
[FAIL] Property animation value 'none'
  assert_equals: expected "none" but got "0s ease 0s 1 normal none running none"
[FAIL] Property animation value 'anim'
  assert_equals: expected "anim" but got "0s ease 0s 1 normal none running anim"
[FAIL] Property animation value 'anim paused both reverse, 4 1s -3s cubic-bezier(0, -2, 1, 3)'
  assert_equals: expected "reverse both paused anim, 1s cubic-bezier(0, -2, 1, 3) -3s 4" but got "0s ease 0s 1 reverse both paused anim, 1s cubic-bezier(0, -2, 1, 3) -3s 4 normal none running none"
[FAIL] Property animation value 'none, none'
  assert_equals: expected "none, none" but got "0s ease 0s 1 normal none running none, 0s ease 0s 1 normal none running none"
[FAIL] Animation with a delay but no duration
  assert_equals: expected "0s 1s" but got "0s ease 1s 1 normal none running none"
Harness: the test ran to completion.
